"Refactor code to shift from store to service pattern"
Code was refactored to shift from using a store pattern to using a service pattern for handling http requests. This included changing the PostsStore and UsersStore to PostService and UserService respectively, which involves renaming the concerned files and updating the references in all components. Additionally, the HttpRequestState was shifted from the store directory to the models directory as it now serves a more generic purpose.

This shift was done to provide better separation of concerns in the codebase, leading to more manageable and readable code. Each service is now responsible for interacting with its corresponding endpoint on the server, while components are now simpler as they only interact with these services to fetch the data they need. This makes the flow of data more predictable and easier to reason about. Finally, the HttpRequestState was made more robust by adding LoadingState, LoadedState, and ErrorState interfaces to better capture the different states an http request can be in.